#36 replace enums with uint/ulong in low-level mechanism params

+++ src/RutokenPkcs11Interop/RutokenPkcs11Interop/HighLevelAPI80/VolumeFormatInfoExtended.cs
@@ -14,8 +14,8 @@ namespace RutokenPkcs11Interop.HighLevelAPI80
             CkVolumeFormatInfoExtended = new CK_VOLUME_FORMAT_INFO_EXTENDED()
             {
                 VolumeSize = volumeSize,
-                AccessMode = accessMode,
-                VolumeOwner = volumeOwner,
+                AccessMode = (ulong) accessMode,
+                VolumeOwner = (ulong) volumeOwner,
                 Flags = flags
             };
         }
